Remove dendrogram
As it depends on bluepy which is not open-sourced.
Also it is not very user friendly as it requires a code snippet prior
to calling the functions in the current API.
Also, it does not work with SONATA circuits.
We will probably do a rework of it when SNAP is open-sourced.

Change-Id: If547de2d417159f818c5aecc31868385a8bbf632