Add an I/O neurondb API
It introduces two classes:

MorphInfo:
A container of all the information regarding a morphology. Information are
what can be found in the neurondb.xml but also lineage information.

MorphologyDB:
The class MorphologyDB can be used to read a 'neurondb.xml' file.

It has public properties to return the data as a:
- Pandas dataframe (MorphologyDB.df)
- Python dictionary (MorphologyDB.data)

It can be used as an iterator that yields MorphInfo objects

It also a 'write()' method to write data to the XML, DAT or CSV format.